use openapi to define API
Previously, methods were mapped to URLs using flask annotations, and
data was validated inside the methods. This is not correct.

Instead, rely on openapi to define the method inputs. Additionally,
use the openapi definition to map URLs to methods.

This commit also handles 500 errors from the inventory service.

NOTE: the returned data from /comparison_report has changed slightly.
The old format for the list of facts was:

```
{UUID1: value1,
 UUID2: value2}
```

It is now:

```
{id: UUID1, value: value1}
{id: UUID2, value: value2}
```

The old style of having data inside of a hash key is not recommended
in the API specs I read.