Handle facts that are only on some systems
Previously, we assumed that all systems have the same set of facts. However,
it's very likely that systems have different sets of facts defined.

This commit adds additional mock facts, and adds a new fact that only appears
if the system UUID starts with a lowercase letter.

Additionally, the comparison now allows for facts that are not defined on all
systems. If a fact is only defined on some systems, the comparison will say
"INCOMPLETE_DATA" as the value for systems where the fact is missing.

Note that if a system really does have a system fact with the value of
"INCOMPLETE_DATA", the comparison report may show the incorrect state. I don't
think there is a way around this, since any value we choose needs to be
serialized to a string for the JSON object.

This commit also creates a `constants.py` which contains all constants. There
were a couple of strings that I did not move into constants (for example,
`"facts"`). It was a bit of a judgment call on which should move and which
shouldn't.