Enable prometheus and gunicorn
Previously, the drift backend app ran with flask's default single-worker
server. This works fine for development, but is not suitable for production.
For example, at startup, the app prints:

```
WARNING: Do not use the development server in a production environment.
Use a production WSGI server instead.
```

This commit enables gunicorn, which is a small webserver that can serve WSGI
apps. The built-in flask server is no longer used.

The best way to run the app now is to invoke `./run_app.sh`. Unit tests can be
run with `run_unit_tests.sh`.

The gunicorn configuration should work out of the box with s2i-python-container.

Additionally, this commit enables Prometheus. We now keep stats on the following:

 * number of exceptions raised when calling `/comparison_report`
 * response time on `/comparison_report` as histogram
 * response time when calling inventory service as histogram

The stats are cleared when the app is restarted.

The health related APIs have been split out into `/mgmt/v0/`. It didn't seem to
make sense to version the health check API in the same way as the application
API. This may or may not make it easier to hide the status check as well, I am
not sure. The new API is defined under `drift/openapi/mgmt_api.spec.yaml`, and
has its own Swagger UI.

Unrelated, this PR also fixes coverage tests that were not running.