Use pipenv instead of requirements.txt
Previously, we kept track of python deps via requirements.txt and
test-requirements.txt. This worked, but did not perform version locking. It
also requires the user to remember to run virtualenv (a step I always forget).

Instead, use pipenv. This creates a Pipfile and Pipfile.lock, so we won't get
an inadvertent library upgrade during a deployment. Additionally, virtualenvs
are handled more cleanly.

This commit migrates to pipenv and also sets a new env var so s2i knows that
pipenv is in use.