use real facts from the inventory service
This commit reads facts off of the "system_profile" fact namespace, so mock
facts do not need to be used anymore.

This commit adds a couple of TODO items in the code. I wanted to get this
commit in so we can get a full end to end of checkin to display, and we can
start treating issues as bugs. I will add cards for the new TODO items.