Do not allow duplicate UUIDs in param list
Previously, users were able to generate a comparison report and specify the
same UUID more than once. This behavior is not defined.

Instead, check that there are no duplicates in the UUID list before rendering
the report. If there are any duplicates, raise an error.

I do not anticipate this being a common issue, so I just check that the UUID
list is the same length as the UUID set. If the list is bigger than the set, we
assume there are duplicates and raise an exception.