call inventory service differently
Previously, we would send inventory service *all* of our UUIDs, and
ask for a paginated result. This resulted in two issues:

 * we were limited by the length of the URL
 * inventory service had to fetch everything every time, calculate
   an offset, and send that back

Instead, we now just ask for 40 UUIDs at a time. This puts less of a
burden on inventory service, removes a URL length limitation, and
overall is cleaner (IMO).

We also track inventory service metrics correctly now.