move x-rh-identity header and increase validation
The `x-rh-identity` header was previously defined as an API parameter.
This caused issues with autogeneration of API bindings.

Instead, move the header's definition to a security scheme.

This PR also enables `strict_validation` and `validate_responses` on
the API. The API appeared to already be in spec and no additional
changes were required besides enabling these additional validators.