do not allow null display_name when copying
For some reason, if an empty display_name is sent via query param,
it's getting passed through openapi to the app, and can create an
empty display_name.

We now validate that the display_name is not empty before we copy a
baseline. This check is only needed for copy and not create, since
create's display_name is validated correctly elsewhere.