Display "N/A" as "" if it's from a baseline
We were previously displaying "N/A" for undefined baseline facts. This
caused a lot of noise on the comparison table since baselines can have
lots of undefined facts.

We now display the empty string for baseline "N/A".

Additionally, we strip out "name" as a fact. This was done before but
regressed.