feature: fetch profiles in time
This commit adds "plumbing" for the calls to fetch profiles in time
from the historical system profiles service. Users can request "pit
ids" and they will be added to the comparison report.

Note that in this commit, we simply bundle historical profiles
alongside regular profiles. These will be broken into two categories
in a future commit.