feature: rename PIT to historical system profile
Previously, we used PIT in some places and historical system profile
in others. This commit gets rid of as many references to PIT as
possible. There are a couple of remaining references in kerlescan that
still need to be removed.

This commit also breaks historical system profiles out as a peer to
`system` and `baseline`. Previously, we lumped them in with `system`s
for proof of concept purposes.