feature: check for "insights" entitlement
Previously, we checked all accounts to see if they had smart
management entitlements.

Instead, check if `ENABLE_SMART_MANAGEMENT_ENTITLEMENT_CHECK` is
enabled first. If so, check for "smart_management" in the
"entitlements" hash.  If it's not enabled, check for "insights".

Once this is disabled in all environments, the flag will go away along
with the smart mgmt entitlement check code.