feature: add listener mode to service
This commit adds a new `SERVICE_MODE` env var. If it is set to
`REST_API`, the service will start as a rest service. This is the
default if `SERVICE_MODE` is not set.

If the `SERVICE_MODE` is set to `LISTENER`, the service will start as a
kafka listener. It can listen for either archival events or delete
events, based on `LISTENER_TYPE`.

This commit moves the code from system-profile-archiver into the
service. The service already has an ORM, and it's cleaner to re-use
this code than to re-implement it as raw SQL in the old archiver.

This commit also removes the ability to POST to create new profiles.
Profiles can only be created via the listener.

Unit test coverage has been moved to 65%. This will be re-raised after
I re-add the profile creation tests.