fix bug when parsing older system profiles
Older system profiles did not have a `captured_date`, so we would
instead return the `created_on` timestamp from the DB. However, we
were returning this as a datetime and not as a string. This resulted
in issues since sometimes `captured_date` would be a datetime, and
sometimes it would be a string.

This commit moves the time conversion code into the model so it's
consistent. It also simplifies the `captured_date` method.