fix: use captured_date as HSP updated field
Previously, we were displaying the "updated" field for HSPs as it comes
from the database. This is of limited value to the user, and we try to
use the "captured_date" instead when possible.

This commit changes the header to use the "captured_date". This makes it
match the date displayed in the dropdown.