fix: return the return value through wrapper
Previously, our transaction wrapper was invoking the correct method,
but was not returning the return value.

This commit adds a return so the value goes through. This is useful
when you want to know the number of deleted records.

Additionally, we added a "LOG_SQL_STATEMENTS" env var. If set, all SQL
will be logged. By default SQL is not logged.