feature: add prometheus metrics server
This commit captures prometheus metrics and exposes them on port 5000.
Right now, it only collects archiver metrics. I'll do deleter metrics
as a different commit after we ensure this one works in CI.

To test: start the archiver and go to http://localhost:5000/ and it
should show a list of metrics. As archiver events come in, the
archiver counters should increase.