[DRFT-176] Add a short_circuit to comparisons
When a system checks in and we create a new hsp, we need to see
if anything has changed, and if so, alert Notifications.
To do this, we need to run a comparison but we only care if any
difference exists at all, not about the whole report, and we
only care about facts present in the baseline.