[DRFT-309] Adds shared secret
When we receive request with x-rh-identity header containing
indentity -> type with value of "System"
we check for "x-rh-drift-internal-api" header value and if it matches
drift shared secret value we skip a call to RBAC service
as we consider this to be a safe call from our own service.

Also projects with kerlescan can use internal_auth_header method to get
the header and use it with API calls to other Drift services.

The secret is read as DRIFT_SHARED_SECRET from ENV.

Adds auth key value to tests instead of "fake key".