[DRFT-582] Fix validation errors for name and fact name
Before, the validation errors for duplicate baseline names and fact names were not matching the mock ups. The sentences were also not in sentence case. This PR properly aligns them with the mock up.

This PR also allows a category and fact to have the same name. But facts cannot share the same name and categories cannot share the same name.