Triggers setting summary when populating the group
This will trigger more than once for every group
but we can do that as long as _set_summary_state method
does not rely on previous run - ie the sets of keys it reads from
and writes into should not overlap.

This way we remove another iteration over grouped_comparisons.
Performance gain depends on the ratio of number of groups
and their size (how many times the function will be called).

Refactors out setting of summary state for grouped comparison