fix: use psycopg2 for ssl FIPS support
In the govcloud environmnet, we have to use FIPS. psycopg2-binary is
compiled without this support, so we have to build it ourselves. These
changes allow for this to happen. The tests are also adjusted to allow
them to complete in a container.

The same changes were made to inventory successfully.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>