feat: calls baselines' update associated systems when system do check-in
ESSNTL-5248

When new kafka message arrives we do a bunch of processings.
When we get answer from bselines service if there are any baselines with
the system (that is represented by the kafka message) we proceed to
update the system in baselines service with newly acquired groups
information.
Then we proceed as normal - do a check for Drift Event.